+++ src/tko/util/decoder.py
@@ -11,16 +11,15 @@ class Decoder:
             encoding = enc_dict["encoding"]
             if encoding is None or enc_dict["confidence"] < 0.90:
                 encoding = "utf-8"
-                
             try:
                 content = raw_data.decode(encoding)
             except UnicodeDecodeError as e:
                 content = raw_data.decode(enc_dict["encoding"])
-            return content
+            return content.replace('\r\n', '\n')
 
     @staticmethod
     def save(file_path: str, content: str):
-        with open(file_path, "w", encoding="utf-8") as file:
+        with open(file_path, "w", encoding="utf-8", newline='\n') as file:
             file.write(content)
 
     @staticmethod